bring back Schlichtkrull subgraph instances
trigger ci
